Add files via upload
"Number Guessing Game" is a Java project hosted on GitHub, designed to entertain and challenge players of all skill levels. With a user-friendly command-line interface, the game generates a random number within a certain range and guides players with hints like "too high" or "too low" to help them guess the correct number. It keeps track of game statistics, making it an ideal choice for those looking to improve their Java skills or simply have fun with a classic guessing game. As an open-source project, it invites contributions and customization, ensuring an engaging and personalized gaming experience.